requirements.txt
folium==0.14.0
pandas==1.5.3
Pillow==9.5.0
plotly==5.14.0
python-dotenv==1.0.0
streamlit==1.19.0
streamlit_folium==0.11.1
